feat: add cache versioning to Metro Bundler and refactor Babel plugin
- Integrated `getCacheVersion` to generate a deterministic cache version for Metro Bundler.
- Added `@crherman7/rechunk-babel-plugin` to Metro configuration for enhanced caching.
- Updated `metro.config.js` to include `config.cacheVersion`.
- Simplified npm scripts in `package.json` by removing redundant `-c` flag.
- Refactored Babel plugin:
  - Moved `getCacheVersion` and `isRechunkDevServerRunning` to `lib/cache.ts`.
  - Removed duplicate logic from `src/index.ts`.
  - Exported cache utilities via `lib/index.ts`.
- Improved TSDoc comments for better developer experience.
