+++ apps/expo/metro.config.js
@@ -1,10 +1,12 @@
 const path = require('path');
+const {cacheVersion} = require('@crherman7/rechunk-babel-plugin');
 const {getDefaultConfig} = require('expo/metro-config');
 
 const root = path.resolve(__dirname, '../..');
 const config = getDefaultConfig(__dirname);
 
 config.watchFolders = [root];
+config.cacheVersion = cacheVersion;
 config.resolver.disableHierarchicalLookup = true;
 config.resolver.nodeModulesPaths = [
   path.resolve(__dirname, 'node_modules'),
