+++ apps/expo/package.json
@@ -4,17 +4,17 @@
   "private": true,
   "main": "index.js",
   "scripts": {
-    "android": "expo start --android -c",
-    "android:dev": "RECHUNK_ENVIRONMENT=dev expo start --android -c",
-    "android:offline": "RECHUNK_ENVIRONMENT=offline expo start --android -c",
-    "ios": "expo start --ios -c",
-    "ios:dev": "RECHUNK_ENVIRONMENT=dev expo start --ios -c",
-    "ios:offline": "RECHUNK_ENVIRONMENT=offline expo start --ios -c",
+    "android": "expo start --android",
+    "android:dev": "RECHUNK_ENVIRONMENT=dev expo start --android",
+    "android:offline": "RECHUNK_ENVIRONMENT=offline expo start --android",
+    "ios": "expo start --ios",
+    "ios:dev": "RECHUNK_ENVIRONMENT=dev expo start --ios",
+    "ios:offline": "RECHUNK_ENVIRONMENT=offline expo start --ios",
     "lint": "pnpm eslint .",
     "lint:fix": "pnpm eslint . --fix",
-    "start": "expo start -c",
-    "start:dev": "RECHUNK_ENVIRONMENT=dev expo start -c",
-    "start:offline": "RECHUNK_ENVIRONMENT=offline expo start -c",
+    "start": "expo start",
+    "start:dev": "RECHUNK_ENVIRONMENT=dev expo start",
+    "start:offline": "RECHUNK_ENVIRONMENT=offline expo start",
     "web": "expo start --web"
   },
   "dependencies": {
