+++ packages/babel-plugin/src/lib/index.ts
@@ -1,2 +1,3 @@
+export * from './cache';
 export * from './ps';
 export * from './workspace';
